requirements.txt
pyautogui>=0.9.54
opencv-python>=4.5.0
Pillow>=8.0.0
pynput>=1.7.0
pyscreeze>=0.1.28
numpy>=1.21.0
flask>=2.0.0
flask-cors>=3.0.10